Some interesting insights
While reading the GORM documentation I saw it was referenced that many2many relationships should be in 2 different files?  This seemed weird...  But I didn't it anyways, hence the explosion of tiny little files in the model directory... I'm not going back... too tired and want to commit

Found the REAL issue I was wrestling with to get Users and Characters linked up properly using this (what I'm not considering to be garbage) ORM and found out how to hack the internal data structures for determining these relations.  That gave birth to the utils directory and the gorm_hacks.go file.  I'll make this generic enough to be used for all many2many relationship changes when I'm not tired and want to go to bed and don't feel like FINALLY getting down to writing the actual logic of this application.

Got some good testing learning under my belt, I'm fairly comfortable with how it operates and with the tests that I've written so far.  Could use a few more assertions on the User/Character relationship and other data objects... now that my trust in GORM has dwindled to nearly nothing

I think that's about it with this changeset...  Yell at me for a longer commit message if it isn't :P.