Brain needs a break
Refactored the database so that I don't have to use the backtick anymore in mysql for anything.  Adopted the 'table names are plural' standard

Test cases are now passing with a few more checks in place to be sure GORM actually does what it says it can do.

Unit tests uncovered that my gorm_hacks was not operating correctly.