AuthenticationCodesRepository Tested
Tests are still not ready for a CI which is disturbing.  I've tried putting them all as subtests under one large test but something is still seeming to execute in parallel and cause database locks which then causes failures.  All unit test functions pass if executed individually on their own... is this the current solution to get travis ci up and running?