Found a potential bug
The roles query wasn't weeding out duplicates.  The version I pasted into the go file was one meant to help understand where things are coming from and it's likely that in a production environment this version will not weed out duplicate roles.  It wouldn't be huge but could lead to some weird log messages in the future.