Lots more to do in the auth-srv
auth-web is functioning ok but have 0 tests to ease my mind.

First course of action is to increase auth-srv to 100% code coverage

Second course of action is to remove ALL REMNANTS OF GORM FROM THE CODE BASE.  THE DB LAYER WILL BE FORGED IN FIRE!  Look into sqlx, I've read good things about it.

Third course of action is to test the auth-web.  With the rest of the code base looking good test wise, I can't have this wart in this set of repositories.