Lots of changes
Found some issue's with GORM/Go when interacting with timestamp columns.  MySQL doesn't like 0001 as a year and it seems that GORM was serioulizing it wrong (had to set the parseTime=true option on the connection).
Go needed me to use pointers for things, specifically all the dates and the corporations allianceId

Removed the daoError in favor of errors.New(...)